Hide PHP 8.1+ test under if
